added sans-serif to font-family to display correctly on Linux
In Firefox on Linux (Ubuntu 22.04, as of writing the most recent LTS release) the theme appeared with serif text. Investigation showed the fonts Lucida and Verdana aren't available on this machine and the default (serif) was picked.

This change adds sans-serif as a last font suggestion to the font-family line.